+++ lib/api/calls/mypage.ts
@@ -1,32 +1,9 @@
 import { build } from '../builder';
-import withTokenInstance from '../instances/withTokenInstance';
+import withRefresh from '../instances/withRefresh';
 
-export const apiProfileUpdateBasic = build<UpdateProfileReq, null>(
-  withTokenInstance,
-  'POST',
-  '/api/mypage/update-basic',
-);
-
-export const apiProfileUpdatePW = build<UpdatePWReq, null>(withTokenInstance, 'POST', '/api/mypage/update-password');
-
-export const apiMyPageBasics = build<Record<string, never>, MyPageBasicInfo>(
-  withTokenInstance,
-  'GET',
-  '/api/mypage/info',
-);
-export const apiMyPageComments = build<{ page: number }, AcdMyComment[]>(
-  withTokenInstance,
-  'GET',
-  '/api/mypage/my-comments',
-);
-export const apiMyPageBookmarks = build<{ page: number }, Course[]>(
-  withTokenInstance,
-  'GET',
-  '/api/mypage/my-bookmarks',
-);
-
-export const apiDeleteAccount = build<{ password: string }, null>(
-  withTokenInstance,
-  'POST',
-  '/api/mypage/delete-profile',
-);
+export const apiProfileUpdateBasic = build<UpdateProfileReq, null>(withRefresh, 'POST', '/api/mypage/update-basic');
+export const apiProfileUpdatePW = build<UpdatePWReq, null>(withRefresh, 'POST', '/api/mypage/update-password');
+export const apiMyPageBasics = build<Record<string, never>, MyPageBasicInfo>(withRefresh, 'GET', '/api/mypage/info');
+export const apiMyPageComments = build<{ page: number }, AcdMyComment[]>(withRefresh, 'GET', '/api/mypage/my-comments');
+export const apiMyPageBookmarks = build<{ page: number }, Course[]>(withRefresh, 'GET', '/api/mypage/my-bookmarks');
+export const apiDeleteAccount = build<{ password: string }, null>(withRefresh, 'POST', '/api/mypage/delete-profile');
